Moved filter and search into distinct folder
